JavaScript File -> Bankist Marketing Website
The functions of the website are primarily done in JavaScript, this is of course a project I've completed through a course. I can take credit for completing it however I must give credit to the creator of the project: Jonas Schmedtmann.

This was an incredible project to understand a lot of modern website functions and uses of JavaScript, almost everything you'll see on this website is useful in real world applications.