JavaScript -> Bankist Marketing Website
This project is being showcased to show my knowledge of a ton of modern JavaScript applications. I love the logic behind a lot of these functions and the thought that went into efficient loading and functionality is top tier.

I must give credit to the creator of this project in which I learned it from although I wrote all of this code it was from a course I was following.

Course Instructor: Jonas Schmedtmann.